Extract FHIR response
