+++ .idea/misc.xml
@@ -7,6 +7,9 @@
         <entry key="app/src/main/res/layout/content_fhir_survey.xml" value="0.15081521739130435" />
         <entry key="app/src/main/res/layout/fhir_survey.xml" value="0.15081521739130435" />
         <entry key="app/src/main/res/layout/questionnaire_layout.xml" value="0.1524822695035461" />
+        <entry key="app/src/main/res/layout/submit_menu.xml" value="0.2578125" />
+        <entry key="app/src/main/res/layout/submit_menu_button.xml" value="0.2578125" />
+        <entry key="app/src/main/res/layout/survey_fragment_layout.xml" value="0.21693840579710144" />
         <entry key="app/src/main/res/layout/survey_layout.xml" value="0.2717391304347826" />
       </map>
     </option>
