+++ app/build.gradle
@@ -83,6 +83,7 @@ dependencies {
     implementation "androidx.compose.ui:ui-tooling-preview:$compose_version"
     implementation "androidx.compose.runtime:runtime-livedata:$compose_version"
     implementation "androidx.compose.material:material-icons-extended:$compose_version"
+    implementation "androidx.compose.ui:ui-viewbinding:$compose_version"
 
     //AndroidX
     implementation 'androidx.core:core-ktx:1.8.0'
