+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".presentation.SurveyActivity">
+            android:name=".presentation.surveys.SurveyActivity">
         </activity>
     </application>
