+++ app/src/main/res/menu/submit_menu.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:id="@+id/submit" android:title="@string/submit" />
+</menu>
